Extracted DeprecatedFeaturesSet to its own file, and rename test suite accordingly
